Bogue tutorials — General instructions

Welcome to the Bogue tutorials!

These tutorials can be simply read online, but of course it's better (and more fun!) to execute the various code chunks by yourself. In fact, all tutorials have been designed so that they can be entirely compiled and executed. Hence, it's a good idea to install the bogue-tutorials package on your machine:

opam install bogue-tutorials --deps-only

This will ensure that the whole code is compatible with your Bogue install. In fact, if you didn't install Bogue beforehand, this will automatically install a compatible version.

How to execute these tutorials

You may simply open an OCaml toplevel (for instance utop), and load Bogue:

#thread;;
#require "bogue";;

Of course we assume that you have a working OCaml environment and installed the Bogue library.

Then, you can simply follow the tutorial by pasting the code chunks into the toplevel.

Another option is to download the source:

git clone https://github.com/sanette/bogue-tutorials.git
cd bogue-tutorials

and execute at once the whole tutorial with the command: make exe from within the tutorial directory. For instance, running the "Hello world" tutorial goes as follows:

cd hello
make exe

If the tutorial opens several Bogue instances, just close the current Bogue window to allow the next one to open.

Ready?

Then let's start:

Hello world

Counter TODO

Self-modifying layouts

More to come...